+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Support for multiple `<code>` tags in shaders.
+- Support for `<code-before>` and `<code-after>` tags in shaders.
 - It is possible to start and stop a loop through the ShaderCanvas instance API.
 - `<draw-loop>` now supports setting the frames per second ("fps" attribute).
 - Show compiler errors for each shader after compilation is done.